Day06 - second part done.

+++ src/day06/race.cpp
@@ -3,7 +3,7 @@
 
 #include "race.hpp"
 
-cRace::cRace(const uint16_t arg_time, const uint16_t arg_distance) :
+cRace::cRace(const uint64_t arg_time, const uint64_t arg_distance) :
 	time(arg_time),
 	distance(arg_distance)
 {
@@ -20,8 +20,8 @@ std::vector<cRace> cRace::Parse(const std::string& arg_path)
 		exit(EXIT_FAILURE);
 	}
 
-	std::vector<uint16_t> times = ParseNumbers(file, "Time:");
-	std::vector<uint16_t> distances = ParseNumbers(file, "Distance:");
+	std::vector<uint64_t> times = ParseNumbers(file, "Time:");
+	std::vector<uint64_t> distances = ParseNumbers(file, "Distance:");
 
 	for (uint8_t i = 0U; i < times.size(); i++)
 	{
@@ -32,9 +32,9 @@ std::vector<cRace> cRace::Parse(const std::string& arg_path)
 	return races;
 }
 
-std::vector<uint16_t> cRace::ParseNumbers(std::ifstream& arg_file, const std::string& arg_prefix)
+std::vector<uint64_t> cRace::ParseNumbers(std::ifstream& arg_file, const std::string& arg_prefix)
 {
-	std::vector<uint16_t> numbers;
+	std::vector<uint64_t> numbers;
 
 	// Get line
 	std::string line;
@@ -43,7 +43,7 @@ std::vector<uint16_t> cRace::ParseNumbers(std::ifstream& arg_file, const std::st
 
 	// Convert string to vector on numbers
 	std::stringstream ss(line);
-	uint16_t number;
+	uint64_t number;
 	while (ss >> number)
 	{
 		numbers.push_back(number);
@@ -64,19 +64,19 @@ uint32_t cRace::GetPowerOfPossibleStrategies(const std::vector<cRace>& arg_races
 	return result;
 }
 
-uint16_t cRace::NumberOfStrategies(void) const
+uint32_t cRace::NumberOfStrategies(void) const
 {
-	uint16_t result = 0U;
+	uint32_t result = 0U;
 
 	// Find center time
-	uint16_t centerWay = BinarySearchStrategy(1U, time - 1U);
+	uint64_t centerWay = BinarySearchStrategy(1U, time - 1U);
 
 	if (centerWay != 0U)
 	{
 		result++;
 
 		// Search for shorter time
-		for (int32_t i = (centerWay - 1U); i >= 1; i--)
+		for (uint64_t i = (centerWay - 1U); i >= 1; i--)
 		{
 			if (!IsStrategyPossible(i))
 			{
@@ -87,7 +87,7 @@ uint16_t cRace::NumberOfStrategies(void) const
 		}
 
 		// Search for longer time
-		for (int32_t i = (centerWay + 1U); i < time; i++)
+		for (uint64_t i = (centerWay + 1U); i < time; i++)
 		{
 			if (!IsStrategyPossible(i))
 			{
@@ -101,27 +101,27 @@ uint16_t cRace::NumberOfStrategies(void) const
 	return result;
 }
 
-uint16_t cRace::BinarySearchStrategy(const uint16_t arg_min, const uint16_t arg_max) const
+uint64_t cRace::BinarySearchStrategy(const uint64_t arg_min, const uint64_t arg_max) const
 {
 	if (arg_min > arg_max)
 	{
 		return 0U;
 	}
 
-	uint16_t center = (arg_max + arg_min) / 2U;
+	uint64_t center = (arg_max + arg_min) / 2U;
 
 	if (IsStrategyPossible(center))
 	{
 		return center;
 	}
 
-	uint16_t shorterTime = BinarySearchStrategy(arg_min, center - 1U);
+	uint64_t shorterTime = BinarySearchStrategy(arg_min, center - 1U);
 	if (shorterTime != 0U)
 	{
 		return shorterTime;
 	}
 
-	uint16_t longerTime = BinarySearchStrategy(center + 1U, arg_max);
+	uint64_t longerTime = BinarySearchStrategy(center + 1U, arg_max);
 	if (longerTime != 0U)
 	{
 		return longerTime;
@@ -130,8 +130,8 @@ uint16_t cRace::BinarySearchStrategy(const uint16_t arg_min, const uint16_t arg_
 	return 0U;
 }
 
-bool cRace::IsStrategyPossible(const uint16_t arg_chargingTime) const
+bool cRace::IsStrategyPossible(const uint64_t arg_chargingTime) const
 {
-	const uint16_t achievedDistance = (time - arg_chargingTime) * arg_chargingTime;
+	const uint64_t achievedDistance = (time - arg_chargingTime) * arg_chargingTime;
 	return achievedDistance > distance;
 }
